More accurate caret movement
When I made block mode work, I saw that the caret is barely missing drawing in the right place.

It was largely reproducible if you move left then right and then try moving right then left. The caret would always settle in a barely (~1-2px) different position depending on which direction the animation was coming from.

The issue was 3 fold:
- We're using `visualPositionToXY` which is less accurate since it spits out `int`s (inside a `Point`) and then casting to double, losing more precision
- We're using `toInt()` directly which always floors, so sometimes we'd be just ever so slightly off to the left
- We're stopping animation before reaching the final spot (`0.01` instead of `0`), so sometimes the animation stops just a bit short of reaching the correct final resting place.

Using `visualPositionToPoint2D` to get the accurate visual position, `round` to be able to round up if needed, and waiting until `dx=0` to finish the animation resolves the inaccuracies.