Use editor caret attributes to determine caret shape
Instead of always forcing one shape or another, we can just check what the IDE has as the caret shape, and make a shape like that.

This makes it so that the plugin is compatible with e.g. Ideavim which likes to switch back and forth between BLOCK and LINE/DEFAULT.

With this change, we can also get rid of the `CaretStyle` settings, since we just follow `CaretVisualAttributes.Shape`.

It seems `SmoothCaretSettings.CaretStyle.BLOCK` is exactly the same as `SmoothCaretSettings.CaretStyle.LINE`, so I instead wrote a change that will try to measure the next character and draw a block caret of that size (or fall back to using 'm' to measure).